CHG: restructure the folders; remove useless functions
