+++ sssd/core/imputers/SSSDS4Imputer.py
@@ -3,7 +3,7 @@ import math
 import torch
 import torch.nn as nn
 
-from sssd.core.imputers.S4Model import S4Layer
+from sssd.core.s4.S4Model import S4Layer
 from sssd.core.utils import calc_diffusion_step_embedding
