+++ sssd/core/imputers/SSSDSAImputer.py
@@ -3,8 +3,8 @@ import torch.nn as nn
 import torch.nn.functional as F
 from einops import rearrange
 
-from sssd.core.imputers.layers.linear import LinearActivation
-from sssd.core.imputers.S4Model import S4
+from sssd.core.layers.linear import LinearActivation
+from sssd.core.s4.S4Model import S4
 from sssd.core.utils import calc_diffusion_step_embedding
