+++ tests/core/layers/test_layers_activation.py
@@ -1,7 +1,7 @@
 import pytest
 import torch.nn as nn
 
-from sssd.core.imputers.layers.activation import Activation
+from sssd.core.layers.activation import Activation
 
 
 def test_activation_identity():
